made memory model, improved

+++ backend_srs/graph_types.h
@@ -3,7 +3,7 @@
 #include <cstdlib>
 #include <vector>
 #include "oper.h"
-using mark_ty = uint64_t;
-using index_ty = uint64_t;
+using mark_ty = size_t;
+using index_ty = size_t;
 using marker_g = std::vector<mark_ty>;
 static constexpr mark_ty nilmark = ~0ULL;
